Flask-Backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
attrs==24.2.0
basketball_reference_scraper==2.0.0
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.2
multidict==6.1.0
nba_api==1.5.0
numpy==1.26.4
outcome==1.3.0.post0
pandas==2.2.2
propcache==0.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
selenium==4.27.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
